Add mime-types for fonts
There is so much bad information available when determining what the correct
mime-types should be. It appears that most browsers currently support
application/x-font-(type) format, but the format that is not fully adopted is
font/(type).

I've commented out the future supported fonts to avoid looking them up again